New pattern type: PredicatePattern.
The PredicatePattern is now what is constructed when a :when token is
seen.  The GuardPattern is now constructed whan a :guard token is
seen.  The difference between the two is that PredicatePattern
predicates are assumed to be comparable (that is, no two
differently-named predicates return true on the same input).  Using
PredicatePattern predicates generates a better decision tree, but will
generate an incorrect decision tree in the event that the predicates
overlap.